edit: multikey search passing tests, limit fix
breaking api change: limit will no longer accept a `end` key, but rather
will work as Arango and god intended, with a `count` key. in effect
though the package claimed to be offering start and end, end was
actually just a misnamed variable, and was passed straight to
ArangoSearch's `COUNT` parameter. This means that the values passed
won't change, but the key name will have to in order to try to adhere to
Arango's specifications. Apologies to those for whom this breaking
change will come with only a patch version change