edit: 40/40 passing tests, multikey is beta
the big drawback is that there are lots of ways for user to input
incorrect objects. Typescript is super helpful, but even still its a lot
of strings and complexity. easy to enter wrong data etc. its part of the
nature of offering cross-collection, cross-analyzer, cross-key boolean
text search. but i believe the middle ground is accessible to most.